prompt tuning typo
