+++ cpeft/prompt_tuning/config.py
@@ -37,37 +37,4 @@ class PromptTuningConfig(PeftConfig):
     def is_prompt_learning(self) -> bool:
         return True
 
-    def save_pretrained(self, save_directory):
-        if os.path.isfile(save_directory):
-            raise ValueError(f"{save_directory} is not a directory.")
 
-        os.makedirs(save_directory, exist_ok=True)
-
-        output_dict = asdict(self)
-        output_path = os.path.join(save_directory, "adapter_config.json")
-
-        with open(output_path, "w") as writer:
-            writer.write(json.dumps(output_dict, indent=2, sort_keys=True))
-
-    @classmethod
-    def from_pretrained(cls, pretrained_model_name_or_path):
-        path = pretrained_model_name_or_path
-
-        config_file = os.path.join(path, "adapter_config.json")
-
-        loaded_attributes = cls.from_json_file(config_file)
-
-        config = PromptTuningConfig()
-
-        for key, value in loaded_attributes.items():
-            if hasattr(config, key):
-                setattr(config, key, value)
-
-        return config
-
-    @classmethod
-    def from_json_file(cls, path_json_file):
-        with open(path_json_file, "r") as file:
-            json_object = json.load(file)
-
-        return json_object
